% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cv_meanKRS_cpp}
\alias{cv_meanKRS_cpp}
\title{Perform cross-validation for meanKRS function based on different lambdas.}
\usage{
cv_meanKRS_cpp(y, x, lambdas, k)
}
\arguments{
\item{y}{numeric vector}

\item{x}{numeric vector}

\item{lambdas}{numeric vector}

\item{k}{int}
}
\value{
Vector of mean squared errors for each lamdba.
}
\description{
Perform cross-validation for meanKRS function based on different lambdas.
}
